requirements.txt
dj-database-url==0.5.0
Django==2.2
django-extensions==2.1.6
django-webpack-loader==0.6.0
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
